+++ apps/api/prisma/migrations/20260830080234_add_account_interest_fields/migration.sql
@@ -1,0 +1,4 @@
+-- AlterTable
+ALTER TABLE "Account" ADD COLUMN     "interestPaymentDay" INTEGER,
+ADD COLUMN     "interestRate" DECIMAL(5,2),
+ADD COLUMN     "taxRate" DECIMAL(5,2);